src/flood_detection_and_photo_storage_cf/requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.24.0
firebase-admin==4.4.0
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0